Refactor proxy param setting
